ActivityManager hangs if process binding takes longer than 10s
On a system with heavy load, when a broadcast receiver timeouts
after the receiving process is started, but before it had time to
bind to ActivityManager, the BroadcastRecord's state is not
reinitialized to IDLE (since finishReceiverLocked() early-returns
as r.receiver is null), so the next call to processNextBroadcast()
will not process the next receiver but instead get stuck on the
current one for which an ANR was already reported.

That state is never updated afterwards: when the process finally
binds via attachApplicationLocked(), mPendingBroadcast is null, so
processCurBroadcastLocked() is not called (and the intent never
sent).

The "hung broadcast" mechanism eventually detects the problem, but
until this is done, all the records in mOrderedBroadcasts are stuck.
For a typical broadcast with 10 receivers for example, this is up to
3 minutes during which ActivityManager is blocking, waiting for
something that will never come. Moreover, the following receivers
of the same broadcast do not get called since the BroadcastRecord
is simply skipped.

This fix makes sure BroadcastRecord.state is always set to IDLE
when a timeout occurs, so that the next receiver in the list will
be processed on the next call to processNextBroadcast(). Since there
were other places where the same pattern was achieved (resetting
members of a BroadcastRecord before scheduleBroadcastsLocked()),
helper methods were added in BroadcastRecord, and some former calls
to finishReceiverLocked() are now using directly the new
BroadcastRecord.finishReceiver().

Change-Id: If4a3f6b608e3bd5205c5e05e686005cfd03647ba