Fix light filter sometimes being slow when changing light conditions.
If the screen is turned off, the light filter is supposed to be reset
when the screen comes back on. On screen off, the light sensor listener
is unregistered, and the code relies on it not being called anymore.
That assumption doesn't hold true, though (verified by checking logs),
which causes the light filter to be filled while the screen is off.