ActivityManagerService: stay away from zombie content providers
When the lowmemorykiller kills a content provider in an unfortunate
moment, an innocent client app can become attached to a zombie provider
and be unconditionally killed moments after, when the dead provider is
being removed.
Prevent this race by addition of isAlive function to Process.java that
determines whether a process is still running and is not only a zombie.
The state of the provider process is checked in getContentProviderImpl.
(includes inc/decProviderCount functions backported from ICS)

Change-Id: I4239f9f15bd0855fa0f3bd8dab0c194bde0add1c