Another Motorola RIL failure workaround.
Sometimes, SETUP_DATA_CALL requests fail with GENERIC_FAILURE error for
some reason. Retrying doesn't help, re-powering the radio does, thus the
workaround attempts to re-power the radio in that scenario.
This workaround likely will also be triggered by BP panics, but this
isn't a problem as the modem will be broken until the reboot anyway.